Proyecto/requirements.txt
Pillow == 9.2.0
requests == 2.28.1
beautifulsoup4 == 4.11.1